Detect file rotation
For systems that provide syscall.Fstat, we can detect
if a file has been replaced by stat'ing it and comparing
its inode with the old file handle. All available data is read
in the old file before moving on to the new file, and a
channel is available for programs to receive notifications
of file rotations.

This feature is not available on windows systems.